Add files via upload
This my line following VEX robotic report involving the robot following a line using a provided VEX robot. The robot lowers reaching arm to a certain height to pick up the payload, and secures it by raising the arm. Following a line and adjusting direction based on the line below it using sensors. Finally dropping off the payload in designated drop off zone.